feat(core): up

+++ packages/core/src/Http/Controllers/Api/BaseModelController.php
@@ -8,10 +8,46 @@ use ContentStash\Core\Http\Resources\ModelResource;
 class BaseModelController extends Controller
 {
     /**
+     * Get policy by model.
+     */
+    public static function getPolicy(string $model)
+    {
+        $policy = str_replace('Models', 'Policies', $model).'Policy';
+        if (! class_exists($policy)) {
+            $policy = 'ContentStash\Core\Policies\BaseModelPolicy';
+        }
+
+        return $policy;
+    }
+
+    /**
+     * Check if the user can perform the action.
+     */
+    public function can(string $model, string $method, $resource = null)
+    {
+        $policy = self::getPolicy($model);
+
+        return app($policy)->$method($model, $resource);
+    }
+
+    /**
+     * Authorize the request.
+     */
+    public function authorize(string $model, string $method, $resource = null)
+    {
+        $can = $this->can($model, $method, $resource);
+
+        if ($can->denied()) {
+            abort(403);
+        }
+    }
+
+    /**
      * Display a listing of the resource.
      */
     public function index(string $model)
     {
+        $this->authorize($model, 'viewAny');
         $model = app($model);
 
         return ModelResource::collection($model->paginate());
@@ -22,6 +58,7 @@ class BaseModelController extends Controller
      */
     public function store(string $model)
     {
+        $this->authorize($model, 'create');
         $model = app($model);
 
         return ModelResource::make($model->create(request()->all()));
@@ -32,6 +69,7 @@ class BaseModelController extends Controller
      */
     public function show(string $model, $resource)
     {
+        $this->authorize($model, 'view', $resource);
         $model = app($model);
 
         return ModelResource::make($model->findOrFail($resource));
@@ -42,10 +80,10 @@ class BaseModelController extends Controller
      */
     public function update(string $model, $resource)
     {
+        $this->authorize($model, 'update', $resource);
         $model = app($model);
 
         $model = $model->findOrFail($resource);
-
         $model->update(request()->all());
 
         return ModelResource::make($model);
@@ -56,10 +94,10 @@ class BaseModelController extends Controller
      */
     public function destroy(string $model, $resource)
     {
+        $this->authorize($model, 'delete', $resource);
+
         $model = app($model);
-
         $model = $model->findOrFail($resource);
-
         $model->delete();
 
         return response()->noContent();
